Module tagua_vm::language

Language manipulation (like the lexer, parser, compiler…).

This module is only responsible of the language manipulation, not its execution. The trivial workflow is the following:

Let datum be of kind &[u8],

The parser module will apply two analyzers and producer:

Lexical analyzer, also called lexer, to basically split datum in a sequence of lexemes, also called tokens,

Syntax analyzer, also called parser, to check that lexemes in the sequence are well ordered,

Finally, if lexemes are well ordered, an Abstract Syntax Tree (AST) is produced.

This AST will be visited by the compiler module visitors to apply transformations or to check properties.

Modules

compiler
Collection of AST transformers and property checkers.

parser
Lexical and syntax analyzers, and AST producer.